Feat : Mypage 기능 테스트 완료

+++ src/main/java/com/supernova/lymming/board/repository/BoardRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface BoardRepository extends JpaRepository<BoardEntity, Integer> {
+public interface BoardRepository extends JpaRepository<BoardEntity, Long> {
 
    Optional<BoardEntity> findByProjectId(Long projectId);
